Make sure 'make' recipes are executed in the right order
Add a couple of rules (with no recipes) to src/Makevars.common
only to describe dependency relationships. This makes sure that
all the recipes in the src/Makevars* files will be executed in
the right order even in the context of parallel execution (e.g.
when doing 'MAKE="make -j4" R CMD INSTALL Rhtslib').