Add C function faidx_fetch_seq_forced_lower() to HTSlib code
This is an alternative to HTSlib C function faidx_fetch_seq(), with the
following difference: if coordinates are outside the actual sequence,
write N's, rather than adjusting the start,end.

Note that this is a modified version of a function of the same name
originally written by Daniel Jones in the seqbias package (the original
function was implemented for an old version of samtools so needed to be
modified to work with htslib-1.7). When the seqbias package will be
migrated to Rhtslib, it will need to use this new version of the function.