Fix kvsprintf() on Windows
Turns out that vsnprintf() is buggy on Windows i.e. it returns -1 when
the output was truncated instead of the number of characters that would
have been written if n had been sufficiently large.
This behavior in turn breaks kvsprintf() on Windows because it assumes
that vsnprintf() behaves the same way on all platforms.

Thanks to Jiefei Wang <szwjf08@gmail.com> for identifying this nasty
kvsprintf() bug on Windows.