New variables to find external packages
When trying to build Rhtslib in spack, the spack built bzip2 and xz
packages could not be found when the internal htslib was built. This PR
adds a couple of environment variables that can be used to define the
locations of bzip2 and xz.

- BZIP2_INCLUDE and XZ_INCLUDE: Can be set like CPPFLAGS,
  '-I/path/to/package/include'
- BZIP2_LIB and XZ_LIB: Can be set like LDFLAGS, `-L/path/to/lib'

A PR for spack that implements this is
https://github.com/spack/spack/pull/14863

This change, or similar, might be useful for other build environments as
well.